Fixes 'wall-rail-gunning' bug.

+++ physics.cpp
@@ -78,8 +78,8 @@ void Physics::update(float dt) {
 }
 
 
-void Physics::rotateGravity(float angle) {
-  gravAngle = int(angle)%360;   // note: '=' not '+=' therefore absolute change
+void Physics::rotateGravity(int angle) {
+  gravAngle = angle%360;   // note: '=' not '+=' therefore absolute change
   float gravityMagnitude = hypot(gravity.x, gravity.y);
   gravity.x = gravityMagnitude * sin(angle*2*M_PI/360);
   gravity.y = gravityMagnitude * cos(angle*2*M_PI/360);
